+++ bigtree/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.2.0"
+__version__ = "1.3.0"
 
 from bigtree._globals import Globals
 from bigtree.binarytree.binarytree import BinaryTree